backend: more endpoints + frontend: technician portal (#6)
* need to test upload obj

* file upload for objects works

* outlined instum setup

* uv setup and obs  file formatting

* generate obj and obs files good

* docker commands

* boilerplate

* reverted to base + included router

* obj_list creation + view works

* ex json

* user_id, test cases

* working .obj upload for objects  + example of maskgen working

* instrument setup

* changed project organization

* git ignore cleanup

* test for uploading .obj files

* cleanup + generate mask endpoint

mask is now stored in db
objs and excluded objs stored
instrument setup stored

* Update backend/tests/test_obj_views.py


* update changes from pull request

* more cleanup

* instrum config setup

* api documentation

* running locally documentation

* tests

* template

* formatting

* First version of frontend

JS9 display, FITS file loading, object file storage

* so much refactoring... project setup completed except for maskgen test

* working tests + machine code endpoints

* changed directories

* cleanup

* API communication

API communication successful, returned body displayed in console, known error that resets JS9

* hours -> deg

* update gitignore

* gitignore

* getting rid of filler code

* updated documentation and gitignore and maskcut endpts

* extra file

* cleanup

* implement feedback

* no more print debug

* Parameters added

all parameters added, ability to view object file details as a table, JS9 disabled (replacement in progress)

* added more endpoints, file organization, and overrides

* updated endpts in template + debugs

* Aladin test added

Js9 still commented out, will be removed

* Update backend/maskgen_api/views.py


* code quality

* Account functionality/Projects

Fully working login/register for accounts through Supabase, added ability to create projects (not connected to API yet), cleanup, aesthetic upgrades, quality of life features...

* validator + db visualizer

* update mask model

* get projects endpoints + get obj lists

* Guest login, code organization

* updated readme + endpoints

* formatting

* more formatting

* test

* technician portal + db changes

* finished technician portal

* divided scroll area and toggle for completed mask

---------

+++ backend/maskgen_api/views.py
@@ -9,7 +9,7 @@ from django.http import FileResponse
 from astropy.table import Table
 
 from .models import Object, Mask, ObjectList, InstrumentConfig, Status, Project, Image
-from .serializers import ObjectSerializer
+from .serializers import ObjectSerializer, MaskSerializer
 from .obs_file_formatting import (
     generate_obj_file,
     generate_obs_file,
@@ -21,7 +21,6 @@ from backend.terminal_helper import run_maskgen, run_maskcut, remove_file
 from .validator import validate
 import json
 import os
-import re
 import io
 import shutil
 
@@ -115,9 +114,16 @@ class InstrumentViewSet(viewsets.ViewSet):
         if version:
             config = get_object_or_404(InstrumentConfig, instrument=pk, version=version)
         else:
-            config = get_object_or_404(
-                InstrumentConfig.objects.order_by("-version"), instrument=pk
+            config = (
+                InstrumentConfig.objects.filter(instrument=pk)
+                .order_by("-version")
+                .first()
             )
+            if config:
+                return Response(
+                    {"error": f"No instrument config found with name '{pk}'"},
+                    status=404,
+                )
 
         return Response(
             {
@@ -209,7 +215,9 @@ class ObjectViewSet(viewsets.ViewSet):
                 status=status.HTTP_400_BAD_REQUEST,
             )
 
-        obj_list = ObjectList.objects.create(name=list_name, user_id=user_id)
+        obj_list = ObjectList.objects.create(
+            name=list_name, user_id=user_id, project_name=proj_name
+        )
 
         for row in data:
             # convert hours to degs if in hrs
@@ -267,8 +275,8 @@ class ObjectViewSet(viewsets.ViewSet):
             {"object lists": list(obj_lists)},
             status=status.HTTP_200_OK,
         )
-
-    @action(detail=False, methods=["delete"], url_path="delete")
+      
+    @action(detail=False, methods=["delete"], url_path="delete_obj")
     def delete_obj(self, request):
         list_name = request.query_params.get("list_name")
         obj_name = request.query_params.get("obj_name")
@@ -279,6 +287,17 @@ class ObjectViewSet(viewsets.ViewSet):
         obj.delete()
         return Response(
             {"message": f"object '{obj_name}' deleted"},
+            status=status.HTTP_200_OK,
+        )
+
+    @action(detail=False, methods=["delete"], url_path="delete_list")
+    def delete_obj_list(self, request):
+        list_name = request.query_params.get("list_name")
+        user_id = request.headers.get("user-id")
+        obj_list = get_object_or_404(ObjectList, name=list_name, user_id=user_id)
+        obj_list.delete()
+        return Response(
+            {"message": f"object list '{list_name}' deleted"},
             status=status.HTTP_200_OK,
         )
 
@@ -465,6 +484,7 @@ class MaskViewSet(viewsets.ViewSet):
             )
             mask = Mask.objects.create(
                 name=filename,
+                user_id=user_id,
                 status=Status.DRAFT,
                 center_ra=data["center_ra"],
                 center_dec=data["center_dec"],
@@ -497,22 +517,6 @@ class MaskViewSet(viewsets.ViewSet):
             )
         else:
             self._file_cleanup(filename)
-            pattern_with_error = r"""
-            (?<=\n)
-            (?:
-                \s*\*\*.*?\*\*\s*\n       # lines starting and ending with "**"
-            )+
-            """
-            matches = re.findall(
-                pattern_with_error, feedback, re.MULTILINE | re.VERBOSE
-            )
-            error_block = matches[0] if matches else ""
-            clean_lines = [line.strip(" *") for line in error_block.splitlines()]
-            cleaned = [s for s in clean_lines if re.search(r"[a-zA-Z]", s)]
-            if len(cleaned) != 0:
-                return False, Response(
-                    {"error": cleaned}, status=status.HTTP_400_BAD_REQUEST
-                )
             return False, Response(
                 {"error": feedback}, status=status.HTTP_400_BAD_REQUEST
             )
@@ -554,6 +558,10 @@ class MaskViewSet(viewsets.ViewSet):
                     generated.append(data["filename"])
                 else:
                     return response
+            return Response(
+                {"created": generated},
+                status=status.HTTP_201_CREATED,
+            )
         elif generate_until_all:
             suffix_count = 1
             excluded_count = 1
@@ -566,12 +574,17 @@ class MaskViewSet(viewsets.ViewSet):
                     return response
                 generated.append(data["filename"])
                 suffix_count += 1
+            return Response(
+                {"created": generated},
+                status=status.HTTP_201_CREATED,
+            )
         else:
-            result, response, _ = self._generate_single_mask(user_id, proj_name, data)
+            result, response, _ = self._generate_single_mask(
+                user_id, proj_name, data, project
+            )
             if result:
                 generated.append(data["filename"])
-            else:
-                return response
+            return response
 
     @action(detail=False, methods=["post"], url_path="finalize")
     def finalize_mask(self, request):
@@ -586,6 +599,46 @@ class MaskViewSet(viewsets.ViewSet):
             mask.save()
             return Response(
                 {"message": "Mask marked as FINALIZED"}, status=status.HTTP_200_OK
+            )
+        else:
+            return Response(
+                {"error": "mask does not exist in this project"},
+                status=status.HTTP_400_BAD_REQUEST,
+            )
+
+    @action(detail=False, methods=["post"], url_path="complete")
+    def complete_mask(self, request):
+        data = request.data
+        proj_name = data["project_name"]
+        user_id = request.headers.get("user-id")
+        mask_name = data["mask_name"]
+        project = Project.objects.get(name=proj_name, user_id=user_id)
+        mask = project.masks.get(name=mask_name)
+        if mask:
+            mask.status = Status.COMPLETED
+            mask.save()
+            return Response(
+                {"message": "Mask marked as COMPLETED"}, status=status.HTTP_200_OK
+            )
+        else:
+            return Response(
+                {"error": "mask does not exist in this project"},
+                status=status.HTTP_400_BAD_REQUEST,
+            )
+
+    @action(detail=False, methods=["post"], url_path="draft")
+    def draft_mask(self, request):
+        data = request.data
+        proj_name = data["project_name"]
+        user_id = request.headers.get("user-id")
+        mask_name = data["mask_name"]
+        project = Project.objects.get(name=proj_name, user_id=user_id)
+        mask = project.masks.get(name=mask_name)
+        if mask:
+            mask.status = Status.DRAFT
+            mask.save()
+            return Response(
+                {"message": "Mask marked as COMPLETED"}, status=status.HTTP_200_OK
             )
         else:
             return Response(
@@ -635,6 +688,24 @@ class MaskViewSet(viewsets.ViewSet):
             status=status.HTTP_200_OK,
         )
 
+    @action(detail=False, methods=["get"], url_path="finalized_masks")
+    def get_finalized_masks(self, request):
+        """
+        Return all masks whose status is FINALIZED
+        """
+        masks = Mask.objects.filter(status="finalized")
+        serializer = MaskSerializer(masks, many=True)
+        return Response(serializer.data)
+
+    @action(detail=False, methods=["get"], url_path="completed_masks")
+    def get_completed_masks(self, request):
+        """
+        Return all masks whose status is COMPLETED
+        """
+        masks = Mask.objects.filter(status="completed")
+        serializer = MaskSerializer(masks, many=True)
+        return Response(serializer.data)
+
 
 class MachineViewSet(viewsets.ViewSet):
     @action(detail=False, methods=["post"], url_path="generate")
@@ -647,6 +718,14 @@ class MachineViewSet(viewsets.ViewSet):
         overwrite = data["overwrite"] == "true"
         project = Project.objects.get(name=proj_name, user_id=user_id)
         mask = project.masks.get(name=mask_name)
+        file_path = f"{PROJECT_DIRECTORY}{API_FOLDER}nc_files/{user_id}/{proj_name}/I{mask_name}.nc"
+
+        if os.path.exists(file_path) and not overwrite:
+            return Response(
+                {"error": "machine code already generated"},
+                status=status.HTTP_400_BAD_REQUEST,
+            )
+
         if mask.status == Status.FINALIZED:
             os.environ["MGPATH"] = MASKGEN_DIRECTORY
             shutil.copy(
@@ -660,11 +739,6 @@ class MachineViewSet(viewsets.ViewSet):
                 f"{MASKGEN_DIRECTORY}/maskcut {mask_name}", overwrite
             )
             if result and "Estimated cutting time" in feedback:
-                remove_file(f"{MASKGEN_DIRECTORY}{mask_name}.SMF")
-                os.makedirs(
-                    os.path.join(f"{PROJECT_DIRECTORY}{API_FOLDER}nc_files", user_id),
-                    exist_ok=True,
-                )
                 os.makedirs(
                     os.path.join(
                         f"{PROJECT_DIRECTORY}{API_FOLDER}nc_files", user_id, proj_name
@@ -711,17 +785,16 @@ class MachineViewSet(viewsets.ViewSet):
                 {"error": "missing required query parameters or headers."},
                 status=status.HTTP_400_BAD_REQUEST,
             )
-        project = Project.objects.get(name=proj_name, user_id=user_id)
-        mask = project.masks.get(name=mask_name)
 
-        if mask.status == Status.FINALIZED:
-            file_path = f"{PROJECT_DIRECTORY}{API_FOLDER}nc_files/{user_id}/{proj_name}/I{mask_name}.nc"
-            file_path = f"{PROJECT_DIRECTORY}{API_FOLDER}nc_files/{user_id}/{proj_name}/I{mask_name}.nc"
+        file_path = f"{PROJECT_DIRECTORY}{API_FOLDER}nc_files/{user_id}/{proj_name}/I{mask_name}.nc"
+
+        if os.path.exists(file_path):
+
             return FileResponse(
                 open(file_path, "rb"), content_type="application/x-netcdf"
             )
         else:
             return Response(
-                {"error": "Mask is not finalized."},
+                {"error": "Machine code has not been generated."},
                 status=status.HTTP_400_BAD_REQUEST,
             )
